+++ base/MAPL_AbstractRegridder.F90
@@ -2,6 +2,7 @@
 
 module MAPL_AbstractRegridderMod
    use MAPL_BaseMod, only: MAPL_UNDEF
+   use MAPL_Constants
    use mapl_RegridderSpec
    use mapl_KeywordEnforcerMod
    use ESMF