+++ base/MAPL_Mod.F90
@@ -13,6 +13,7 @@ module MAPL_Mod
   use MAPL_CommsMod
   use MAPL_GenericMod
   use MAPL_VarSpecMod
+! For temporary backward compatibility after Constants Library
   use MAPL_ConstantsMod
   use MAPL_ConfigMod
   use MAPL_SortMod